# src/main/gen/src/main/java/io/security/portfolio/api/TeamsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.security.portfolio.api;

import io.security.portfolio.dto.AddTeamMember201Response;
import io.security.portfolio.dto.CreateTeam201Response;
import io.security.portfolio.dto.ListTeamMembers200Response;
import io.security.portfolio.dto.ListTeams200Response;
import io.security.portfolio.dto.ModelApiResponse;
import io.security.portfolio.dto.TeamMemberRequest;
import io.security.portfolio.dto.TeamRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-07T12:21:57.339446635+09:00[Asia/Seoul]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "Teams", description = "the Teams API")
public interface TeamsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/teams/{teamId}/members : 팀 멤버 추가
     *
     * @param teamId  (required)
     * @param teamMemberRequest  (required)
     * @return 추가된 팀 멤버 정보 (status code 201)
     */
    @Operation(
        operationId = "addTeamMember",
        summary = "팀 멤버 추가",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "201", description = "추가된 팀 멤버 정보", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AddTeamMember201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/teams/{teamId}/members",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AddTeamMember201Response> addTeamMember(
        @Parameter(name = "teamId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("teamId") Long teamId,
        @Parameter(name = "TeamMemberRequest", description = "", required = true) @Valid @RequestBody TeamMemberRequest teamMemberRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"roleInTeam\" : \"roleInTeam\", \"teamId\" : 0, \"userId\" : 6 }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/teams : 팀 생성
     *
     * @param teamRequest  (required)
     * @return 생성된 팀 정보 (status code 201)
     */
    @Operation(
        operationId = "createTeam",
        summary = "팀 생성",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "201", description = "생성된 팀 정보", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateTeam201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/teams",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<CreateTeam201Response> createTeam(
        @Parameter(name = "TeamRequest", description = "", required = true) @Valid @RequestBody TeamRequest teamRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/teams/{teamId} : 팀 상세 조회
     *
     * @param teamId  (required)
     * @return 팀 상세 정보 (status code 200)
     */
    @Operation(
        operationId = "getTeamById",
        summary = "팀 상세 조회",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "200", description = "팀 상세 정보", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateTeam201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/teams/{teamId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CreateTeam201Response> getTeamById(
        @Parameter(name = "teamId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("teamId") Long teamId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/teams/{teamId}/members : 팀 멤버 목록 조회
     *
     * @param teamId  (required)
     * @return 팀 멤버 목록 (status code 200)
     */
    @Operation(
        operationId = "listTeamMembers",
        summary = "팀 멤버 목록 조회",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "200", description = "팀 멤버 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListTeamMembers200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/teams/{teamId}/members",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListTeamMembers200Response> listTeamMembers(
        @Parameter(name = "teamId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("teamId") Long teamId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"roleInTeam\" : \"roleInTeam\", \"teamId\" : 0, \"userId\" : 6 }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"roleInTeam\" : \"roleInTeam\", \"teamId\" : 0, \"userId\" : 6 } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/teams : 팀 목록 조회
     *
     * @param name 팀 이름 검색용 (선택) (optional)
     * @return 팀 목록 (status code 200)
     */
    @Operation(
        operationId = "listTeams",
        summary = "팀 목록 조회",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "200", description = "팀 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListTeams200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/teams",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListTeams200Response> listTeams(
        @Parameter(name = "name", description = "팀 이름 검색용 (선택)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "name", required = false) String name
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/teams/{teamId}/members/{userId} : 팀 멤버 제거
     *
     * @param teamId  (required)
     * @param userId  (required)
     * @return 삭제 결과 (status code 200)
     */
    @Operation(
        operationId = "removeTeamMember",
        summary = "팀 멤버 제거",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "200", description = "삭제 결과", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/teams/{teamId}/members/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ModelApiResponse> removeTeamMember(
        @Parameter(name = "teamId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("teamId") Long teamId,
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") Long userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : \"\", \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /api/teams/{teamId}/members/{userId} : 팀 멤버 역할 변경
     *
     * @param teamId  (required)
     * @param userId  (required)
     * @param teamMemberRequest  (required)
     * @return 수정된 팀 멤버 정보 (status code 200)
     */
    @Operation(
        operationId = "updateTeamMember",
        summary = "팀 멤버 역할 변경",
        tags = { "Teams" },
        responses = {
            @ApiResponse(responseCode = "200", description = "수정된 팀 멤버 정보", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AddTeamMember201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/api/teams/{teamId}/members/{userId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AddTeamMember201Response> updateTeamMember(
        @Parameter(name = "teamId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("teamId") Long teamId,
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") Long userId,
        @Parameter(name = "TeamMemberRequest", description = "", required = true) @Valid @RequestBody TeamMemberRequest teamMemberRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"roleInTeam\" : \"roleInTeam\", \"teamId\" : 0, \"userId\" : 6 }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}