# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Document Authorization API
  version: 1.0.0
  description: |
    RBAC + Team Role + ABAC + ACL(PBAC 기반) 구조를 가진 문서 인가 데모 API.
    - 글로벌 RBAC: users / roles / user_roles
    - 팀/조직: teams / team_members
    - 문서: documents
    - ACL(문서별 권한): document_permissions
    - 감사 로그: audit_logs
    - 인증: refresh_tokens
servers:
  - url: http://localhost:8080
    description: Local dev server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        message:
          type: string
          nullable: true
          description: 추가 메시지 (없으면 null)
        data:
          nullable: true
          description: 실제 응답 데이터 (엔드포인트별로 타입 상이)

    # =====================
    # Auth
    # =====================
...
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    TeamMemberResponse:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        roleInTeam:
          type: string
          description: "OWNER, MANAGER, MEMBER"
        createdAt:
          type: string
          format: date-time
    TeamMemberRequest:
      type: object
      required: [userId, roleInTeam]
      properties:
        userId:
          type: integer
          format: int64
        roleInTeam:
          type: string
          description: "OWNER, MANAGER, MEMBER"

    # =====================
    # Document
    # =====================
...
paths:
...
  /api/teams:
    get:
      tags: [Teams]
      summary: 팀 목록 조회
      operationId: listTeams
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: 팀 이름 검색용 (선택)
      responses:
        '200':
          description: 팀 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamResponse'
    post:
      tags: [Teams]
      summary: 팀 생성
      operationId: createTeam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
      responses:
        '201':
          description: 생성된 팀 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TeamResponse'
  /api/teams/{teamId}:
    get:
      tags: [Teams]
      summary: 팀 상세 조회
      operationId: getTeamById
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 팀 상세 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TeamResponse'
  /api/teams/{teamId}/members:
    get:
      tags: [Teams]
      summary: 팀 멤버 목록 조회
      operationId: listTeamMembers
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 팀 멤버 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamMemberResponse'
    post:
      tags: [Teams]
      summary: 팀 멤버 추가
      operationId: addTeamMember
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberRequest'
      responses:
        '201':
          description: 추가된 팀 멤버 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TeamMemberResponse'
  /api/teams/{teamId}/members/{userId}:
    patch:
      tags: [Teams]
      summary: 팀 멤버 역할 변경
      operationId: updateTeamMember
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberRequest'
      responses:
        '200':
          description: 수정된 팀 멤버 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TeamMemberResponse'
    delete:
      tags: [Teams]
      summary: 팀 멤버 제거
      operationId: removeTeamMember
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 삭제 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # =====================
  # Documents
  # =====================
...